Allow macro expansion (#38)
In order to get the expanded version of the crate, we need to run Cargo.
The problem is that the main function already sets up a cargo instance
and that we can't run two instances of it at the same time (a lockfile
is used by Cargo in order to prevent this).

In order to allow to run Cargo, some small changes were necessary. More
precisely, we need to get the crate code before running any cargo
instance, but we can't perform any analysis on it since cargo is not
running. As such, a new step has been added in the main function, where
we just *extract* the crate code and store it in a temporary data
structure. This data structure is then passed to `input::read_file`,
which has been renamed to `input::read_code` in order to better express
the fact that files don't exist at this point.

There's a small subtility: the expanded code generated by rustc has a
prelude import manually added at the beginning of the crate. The
problem is that we're emitting a warning for every glob import,
including this prelude call. In order to fix this issue, a filter has
been added before signaling that we have met glob import, in order
to not report the prelude import as a glob import.

Co-authored-by: Dominic <git@msrd0.de>